--- Migrator.API/Migrator.API.csproj
@@ -1,18 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>net8.0</TargetFramework>
-    <Nullable>enable</Nullable>
-    <ImplicitUsings>enable</ImplicitUsings>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="6.6.2" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Migrator.Application\Migrator.Application.csproj" />
-    <ProjectReference Include="..\Migrator.Infrastructure\Migrator.Infrastructure.csproj" />
-  </ItemGroup>
-
-</Project>
