--- Migrator.Application/Dto/DuplicteResponse.cs
@@ -1,8 +1,0 @@
-﻿namespace Migrator.Application.Dto
-{
-    public class DuplicteResponse
-    {
-        public string Name { get; set; }
-        public int Count { get; set; }
-    }
-}
