--- Migrator.Application/Dto/OrganisationDto.cs
@@ -1,11 +1,0 @@
-﻿namespace Migrator.Application.Dto
-{
-    internal class OrganisationDto
-    {
-        public string Name { get; set; }
-        public string TownCity { get; set; }
-        public string County { get; set; }
-        public string TypeAndRating { get; set; }
-        public string Route { get; set; }
-    }
-}
