--- Migrator.Core/Enums/RouteEnum.cs
@@ -1,23 +1,0 @@
-﻿namespace Migrator.Core.Enums
-{
-    public enum RouteEnum : int
-    {
-        CharityWorker,
-        CreativeWorker,
-        GlobalBusinessMobility_GraduateTrainee,
-        GlobalBusinessMobility_SecondmentWorker,
-        GlobalBusinessMobility_SeniorOrSpecialistWorker,
-        GlobalBusinessMobility_ServiceSupplier,
-        GlobalBusinessMobility_UKExpansionWorker,
-        GovernmentAuthorisedExchange,
-        InternationalAgreement,
-        InternationalSportsperson,
-        IntraCompanyTransfers_ICT,
-        IntraCompanyRoutes,
-        ReligiousWorker,
-        ScaleUp,
-        SeasonalWorker,
-        SkilledWorker,
-        Tier2MinistersOfReligion
-    }
-}
