+++ VisaSponsorshipScout.API/VisaSponsorshipScout.API.csproj
@@ -1,0 +1,18 @@
+﻿<Project Sdk="Microsoft.NET.Sdk.Web">
+
+  <PropertyGroup>
+    <TargetFramework>net8.0</TargetFramework>
+    <Nullable>enable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Swashbuckle.AspNetCore" Version="6.6.2" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\VisaSponsorshipScout.Application\VisaSponsorshipScout.Application.csproj" />
+    <ProjectReference Include="..\VisaSponsorshipScout.Infrastructure\VisaSponsorshipScout.Infrastructure.csproj" />
+  </ItemGroup>
+
+</Project>
