+++ VisaSponsorshipScout.Application/Dto/DuplicteResponse.cs
@@ -1,0 +1,8 @@
+﻿namespace VisaSponsorshipScout.Application.Dto
+{
+    public class DuplicteResponse
+    {
+        public string Name { get; set; }
+        public int Count { get; set; }
+    }
+}
