+++ VisaSponsorshipScout.Core/Enums/RouteEnum.cs
@@ -1,0 +1,23 @@
+﻿namespace VisaSponsorshipScout.Core.Enums
+{
+    public enum RouteEnum : int
+    {
+        CharityWorker,
+        CreativeWorker,
+        GlobalBusinessMobility_GraduateTrainee,
+        GlobalBusinessMobility_SecondmentWorker,
+        GlobalBusinessMobility_SeniorOrSpecialistWorker,
+        GlobalBusinessMobility_ServiceSupplier,
+        GlobalBusinessMobility_UKExpansionWorker,
+        GovernmentAuthorisedExchange,
+        InternationalAgreement,
+        InternationalSportsperson,
+        IntraCompanyTransfers_ICT,
+        IntraCompanyRoutes,
+        ReligiousWorker,
+        ScaleUp,
+        SeasonalWorker,
+        SkilledWorker,
+        Tier2MinistersOfReligion
+    }
+}
