+++ VisaSponsorshipScout.Core/VisaSponsorshipScout.Core.csproj
