+++ VisaSponsorshipScout.sln
@@ -1,0 +1,43 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.12.35707.178
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VisaSponsorshipScout.API", "VisaSponsorshipScout.API\VisaSponsorshipScout.API.csproj", "{B96C60E4-1CC4-4AF4-9482-8D42FC76D0EC}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VisaSponsorshipScout.Core", "VisaSponsorshipScout.Core\VisaSponsorshipScout.Core.csproj", "{5D8F03F1-7689-40D9-99ED-01A694C92801}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VisaSponsorshipScout.Application", "VisaSponsorshipScout.Application\VisaSponsorshipScout.Application.csproj", "{E564FA65-C411-44AB-961D-3E7CF3BFD6A9}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "VisaSponsorshipScout.Infrastructure", "VisaSponsorshipScout.Infrastructure\VisaSponsorshipScout.Infrastructure.csproj", "{28F82DE5-4DB8-4B5B-9F65-535190638288}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{B96C60E4-1CC4-4AF4-9482-8D42FC76D0EC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B96C60E4-1CC4-4AF4-9482-8D42FC76D0EC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B96C60E4-1CC4-4AF4-9482-8D42FC76D0EC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B96C60E4-1CC4-4AF4-9482-8D42FC76D0EC}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5D8F03F1-7689-40D9-99ED-01A694C92801}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5D8F03F1-7689-40D9-99ED-01A694C92801}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5D8F03F1-7689-40D9-99ED-01A694C92801}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5D8F03F1-7689-40D9-99ED-01A694C92801}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E564FA65-C411-44AB-961D-3E7CF3BFD6A9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E564FA65-C411-44AB-961D-3E7CF3BFD6A9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E564FA65-C411-44AB-961D-3E7CF3BFD6A9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E564FA65-C411-44AB-961D-3E7CF3BFD6A9}.Release|Any CPU.Build.0 = Release|Any CPU
+		{28F82DE5-4DB8-4B5B-9F65-535190638288}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{28F82DE5-4DB8-4B5B-9F65-535190638288}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{28F82DE5-4DB8-4B5B-9F65-535190638288}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{28F82DE5-4DB8-4B5B-9F65-535190638288}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {FB1BADD4-799D-432D-BE9A-378308A3F99B}
+	EndGlobalSection
+EndGlobal
